refactor: モデルを音声合成モジュール下へ移動
