+++ test/tts_pipeline/test_wave_synthesizer.py
@@ -2,7 +2,8 @@
 
 import numpy as np
 
-from voicevox_engine.model import AccentPhrase, AudioQuery, Mora
+from voicevox_engine.model import AudioQuery
+from voicevox_engine.tts_pipeline.model import AccentPhrase, Mora
 from voicevox_engine.tts_pipeline.tts_engine import (
     apply_intonation_scale,
     apply_output_sampling_rate,
